requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
flask-restx==1.3.0
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.0
pluggy==1.5.0
pytest==8.2.1
python-dateutil==2.9.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
tomli==2.0.1
Werkzeug==3.0.3
